test(`MemoryDao`): 테스트용 Fake 객체 추가 및 서비스 테스트에 반영
