+++ src/main/java/wooteco/subway/exception/constant/NotExistException.java
@@ -1,6 +1,6 @@
 package wooteco.subway.exception.constant;
 
-public class NotExistException extends IllegalArgumentException{
+public class NotExistException extends IllegalArgumentException {
 
     private static final String MESSAGE = "존재하지 않습니다.";
