+++ src/main/resources/application.yml
@@ -1,5 +1,5 @@
 spring:
-  datasource:           # datasource 설정
+  datasource: # datasource 설정
     driver-class-name: org.h2.Driver
     url: jdbc:h2:mem:testdb
     username: sa
