+++ src/test/java/wooteco/subway/application/StationServiceTest.java
@@ -5,9 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
-import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.dao.MemoryStationDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.constant.BlankArgumentException;
@@ -17,11 +15,7 @@ import wooteco.subway.exception.constant.NotExistException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-@JdbcTest
 class StationServiceTest {
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
 
     private StationDao stationDao;
 
@@ -29,7 +23,7 @@ class StationServiceTest {
 
     @BeforeEach
     void setUp() {
-        stationDao = new StationDao(jdbcTemplate);
+        stationDao = new MemoryStationDao();
         stationService = new StationService(stationDao);
     }
